backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
psycopg[binary]==3.2.3
django-cors-headers==4.6.0
python-decouple==3.8
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.9.0
openpyxl==3.1.5

